REFACTOR - 共通処理部分を切り出し、mixinで読み込み共通化 共通化したため、フォルダも同じ場所に統一化
