+++ src/components/molecules/CategoriesList/CategoriesListBase.js
@@ -1,0 +1,17 @@
+export default {
+  props: {
+    category: {
+      type: Object,
+      required: true
+    },
+    value: {
+      type: String,
+      required: true
+    }
+  },
+  methods: {
+    emitMethod(value) {
+      return this.$emit("change", value);
+    }
+  }
+};
