+++ src/components/molecules/CategoriesList/ExpansionState.js
